requirements.txt
Django==4.0.1
openbb-nightly
djangorestframework
drf-yasg